+++ ubo_app/services/050-lightdm/reducer.py
@@ -5,27 +5,27 @@ from dataclasses import replace
 
 from redux import InitAction, InitializationActionError
 
-from ubo_app.store.services.ssh import (
-    SSHAction,
-    SSHClearEnabledStateAction,
-    SSHState,
-    SSHUpdateStateAction,
+from ubo_app.store.services.lightdm import (
+    LightDMAction,
+    LightDMClearEnabledStateAction,
+    LightDMState,
+    LightDMUpdateStateAction,
 )
 
 
 def reducer(
-    state: SSHState | None,
-    action: SSHAction | InitAction,
-) -> SSHState:
+    state: LightDMState | None,
+    action: LightDMAction | InitAction,
+) -> LightDMState:
     if state is None:
         if isinstance(action, InitAction):
-            return SSHState(is_active=False, is_enabled=False)
+            return LightDMState(is_active=False, is_enabled=False)
         raise InitializationActionError(action)
 
-    if isinstance(action, SSHClearEnabledStateAction):
+    if isinstance(action, LightDMClearEnabledStateAction):
         return replace(state, is_enabled=None)
 
-    if isinstance(action, SSHUpdateStateAction):
+    if isinstance(action, LightDMUpdateStateAction):
         if action.is_active is not None:
             state = replace(state, is_active=action.is_active)
         if action.is_enabled is not None: