+++ ubo_app/system/system_manager/main.py
@@ -12,10 +12,11 @@ from pathlib import Path
 from threading import Thread
 
 from ubo_app.constants import USERNAME
+from ubo_app.error_handlers import setup_error_handling
 from ubo_app.logging import add_file_handler, add_stdout_handler, get_logger
 from ubo_app.system.system_manager.docker import docker_handler
 from ubo_app.system.system_manager.led import LEDManager
-from ubo_app.system.system_manager.ssh import ssh_handler
+from ubo_app.system.system_manager.service_manager import system_handler
 
 SOCKET_PATH = Path(os.environ.get('RUNTIME_DIRECTORY', '/run/ubo')).joinpath(
     'system_manager.sock',
@@ -36,13 +37,14 @@ def handle_command(command: str) -> str | None:
     elif header == 'docker':
         thread = Thread(target=docker_handler, args=(incoming[0],))
         thread.start()
-    elif header == 'ssh':
-        return ssh_handler(incoming[0])
+    elif header == 'service':
+        return system_handler(incoming[0], incoming[1])
     return None
 
 
 def main() -> None:
     """Initialise the System-Manager."""
+    setup_error_handling()
     logger.debug('Initialising System-Manager...')
 
     led_manager.run_command_thread_safe('spinning_wheel 255 255 255 50 6 100'.split())