+++ ubo_app/utils/monitor_unit.py
@@ -2,6 +2,8 @@
 # ruff: noqa: D100, D101, D102, D103, D104, D105, D107
 from __future__ import annotations
 
+import asyncio
+import subprocess
 from typing import Callable
 
 from sdbus import DbusInterfaceCommonAsync, dbus_property_async
@@ -30,6 +32,7 @@ def to_dbus_string(string: str) -> str:
 
 
 async def monitor_unit(unit_name: str, callback: Callable[[str], None]) -> None:
+    """Monitor the active state of a systemd unit."""
     bus = get_system_bus()
     system_service = SystemdUnitInterface.new_proxy(
         bus=bus,
@@ -40,3 +43,29 @@ async def monitor_unit(unit_name: str, callback: Callable[[str], None]) -> None:
     async for _ in system_service.properties_changed:
         active_state = await system_service.active_state
         callback(active_state)
+
+
+async def is_unit_active(unit: str) -> bool:
+    """Check if the systemd unit is active."""
+    process = await asyncio.create_subprocess_exec(
+        '/usr/bin/env',
+        'systemctl',
+        'is-active',
+        unit,
+        stdout=subprocess.PIPE,
+    )
+    stdout, _ = await process.communicate()
+    return stdout.strip() == b'active'
+
+
+async def is_unit_enabled(unit: str) -> bool:
+    """Check if the systemd unit is enabled."""
+    process = await asyncio.create_subprocess_exec(
+        '/usr/bin/env',
+        'systemctl',
+        'is-enabled',
+        unit,
+        stdout=subprocess.PIPE,
+    )
+    stdout, _ = await process.communicate()
+    return stdout.strip() == b'enabled'